Demo v.8 update
1. Added key system. You can now specify what items/keys are required to
be picked up in order to unclock  something. In the Demo specifically,
you need to collect two keys in order to be able to open doors with the
door botton.
2, Enable all parts of the rope to be able to interact with the
Elemental. Only the end of the rope was enabled previously.
3. Changed inputs for rope grabbing. You can now grab to the end of a
rope by pressing E once and release it by pressing E again. No longer
need to hold E.